update sdk to 0.0.7 version
