+++ sdk/rust/src/allocator/mod.rs
@@ -1,30 +1,24 @@
-use std::alloc::{Alloc, Global, Layout};
+use std::num::NonZeroUsize;
+use std::ptr::NonNull;
+use crate::memory::{alloc, dealloc};
 
-/// Allocates a memory region of given size and returns its address. Actually is
-/// just a wrapper for [`GlobalAlloc::alloc`].
-///
-/// # Safety
-///
-/// See [`GlobalAlloc::alloc`].
-///
-/// [`GlobalAlloc::alloc`]: https://doc.rust-lang.org/core/alloc/trait.GlobalAlloc.html#tymethod.alloc
-///
-pub unsafe fn alloc(size: NonZeroUsize) -> MemResult<NonNull<u8>> {
-    let layout: Layout = Layout::from_size_align(size.get(), mem::align_of::<u8>())?;
-    Global.alloc(layout).map_err(Into::into)
+/// Allocates memory area of specified size and returns its address.
+/// Used from the host environment for memory allocation while parameters passing.
+#[no_mangle]
+pub unsafe fn allocate(size: usize) -> NonNull<u8> {
+    let non_zero_size = NonZeroUsize::new(size)
+        .unwrap_or_else(|| panic!("[Error] Allocation of zero bytes is not allowed."));
+    alloc(non_zero_size)
+        .unwrap_or_else(|_| panic!("[Error] Allocation of {} bytes failed.", size))
 }
 
-/// Deallocates a memory region for current pointer and size. Actually is
-/// just a wrapper for [`GlobalAlloc::dealloc`].
-///
-/// # Safety
-///
-/// See [`GlobalAlloc::dealloc`].
-///
-/// [`GlobalAlloc::dealloc`]: https://doc.rust-lang.org/core/alloc/trait.GlobalAlloc.html#tymethod.dealloc
-///
-pub unsafe fn dealloc(ptr: NonNull<u8>, size: NonZeroUsize) -> MemResult<()> {
-    let layout = Layout::from_size_align(size.get(), mem::align_of::<u8>())?;
-    Global.dealloc(ptr, layout);
-    Ok(())
+/// Deallocates memory area for provided memory pointer and size.
+/// Used from the host environment for memory deallocation after reading results
+/// of function from Wasm memory.
+#[no_mangle]
+pub unsafe fn deallocate(ptr: NonNull<u8>, size: usize) {
+    let non_zero_size = NonZeroUsize::new(size)
+        .unwrap_or_else(|| panic!("[Error] Deallocation of zero bytes is not allowed."));
+    dealloc(ptr, non_zero_size)
+        .unwrap_or_else(|_| panic!("[Error] Deallocate failed for ptr={:?} size={}.", ptr, size));
 }
