+++ vm/examples/counter/src/lib.rs
@@ -16,17 +16,7 @@
 
 //! Wrapper for counter (a test for Fluence network).
 //!
-//! Provides the public method `invoke` for increment a counter and get its current state and some
-//! public service methods (`allocation`, `deallocation`).
-
-#![feature(allocator_api)]
-#![feature(alloc)]
-extern crate alloc;
-extern crate core;
-extern crate fluence_sdk as fluence;
-use std::num::NonZeroUsize;
-
-use std::ptr::NonNull;
+//! Provides the public method `invoke` for increment a counter and get its current state.
 
 mod counter;
 
@@ -45,25 +35,4 @@ pub unsafe fn invoke(_ptr: *mut u8, _len: usize) -> usize {
             panic!("[Error] Putting the result string into a raw memory was failed")
         })
         .as_ptr() as usize
-}
-
-/// Allocates memory area of specified size and returns its address.
-/// Used from the host environment for memory allocation while parameters passing.
-#[no_mangle]
-pub unsafe fn allocate(size: usize) -> NonNull<u8> {
-    let non_zero_size = NonZeroUsize::new(size)
-        .unwrap_or_else(|| panic!("[Error] Allocation of zero bytes is not allowed."));
-    fluence::memory::alloc(non_zero_size)
-        .unwrap_or_else(|_| panic!("[Error] Allocation of {} bytes failed.", size))
-}
-
-/// Deallocates memory area for provided memory pointer and size.
-/// Used from the host environment for memory deallocation after reading results
-/// of function from Wasm memory.
-#[no_mangle]
-pub unsafe fn deallocate(ptr: NonNull<u8>, size: usize) {
-    let non_zero_size = NonZeroUsize::new(size)
-        .unwrap_or_else(|| panic!("[Error] Deallocation of zero bytes is not allowed."));
-    fluence::memory::dealloc(ptr, non_zero_size)
-        .unwrap_or_else(|_| panic!("[Error] Deallocate failed for ptr={:?} size={}.", ptr, size));
 }
