+++ vm/examples/fleth/src/lib.rs
@@ -20,10 +20,8 @@
 #![allow(dead_code)]
 
 use log::{info, warn, error};
-use fluence_sdk as fluence;
 
 use std::error::Error;
-use std::num::NonZeroUsize;
 use std::ptr::NonNull;
 
 /// Result for all possible Error types.
@@ -58,39 +56,6 @@ pub unsafe fn invoke(ptr: *mut u8, len: usize) -> NonNull<u8> {
     // return pointer to result in memory
     fluence::memory::write_str_to_mem("OK")
         .unwrap_or_else(|_| log_and_panic("Putting result string to the memory was failed.".into()))
-}
-
-//
-// Public functions for memory management
-//
-
-/// Allocates memory area of specified size and returns its address.
-/// Used from the host environment for memory allocation while parameters passing.
-#[no_mangle]
-pub unsafe fn allocate(size: usize) -> NonNull<u8> {
-    info!("allocate starts with size={}", size);
-    let non_zero_size = NonZeroUsize::new(size).unwrap_or_else(|| {
-        log_and_panic("[Error] Allocation of zero bytes is not allowed.".into())
-    });
-    fluence::memory::alloc(non_zero_size)
-        .unwrap_or_else(|_| log_and_panic(format!("[Error] Allocation of {} bytes failed.", size)))
-}
-
-/// Deallocates memory area for provided memory pointer and size.
-/// Used from the host environment for memory deallocation after reading results
-/// of function from Wasm memory.
-#[no_mangle]
-pub unsafe fn deallocate(ptr: NonNull<u8>, size: usize) {
-    info!("deallocate starts with ptr={:?}, size={}", ptr, size);
-    let non_zero_size = NonZeroUsize::new(size).unwrap_or_else(|| {
-        log_and_panic("[Error] Deallocation of zero bytes is not allowed.".into())
-    });
-    fluence::memory::dealloc(ptr, non_zero_size).unwrap_or_else(|_| {
-        log_and_panic(format!(
-            "[Error] Deallocate failed for ptr={:?} size={}.",
-            ptr, size
-        ))
-    });
 }
 
 fn log_and_panic(msg: String) -> ! {
