+++ vm/examples/hello-user/src/lib.rs
@@ -18,10 +18,10 @@
 
 #![feature(allocator_api)]
 #![feature(alloc)]
-use fluence_sdk as fluence;
-use log::{error, info, warn};
 
-use std::num::NonZeroUsize;
+use fluence;
+
+use log::{error, info, warn};
 use std::ptr::NonNull;
 
 /// Initializes `WasmLogger` instance and returns a pointer to error message as a string
@@ -48,27 +48,6 @@ pub unsafe fn invoke(ptr: *mut u8, len: usize) -> NonNull<u8> {
     // return a pointer to the result in memory
     fluence::memory::write_str_to_mem(format!("Hello from Fluence to {}", user_name).as_str())
         .unwrap_or_else(|_| log_and_panic("Putting result string to the memory was failed.".into()))
-}
-
-/// Allocates memory area of specified size and returns its address.
-/// Used from the host environment for memory allocation while parameters passing.
-#[no_mangle]
-pub unsafe fn allocate(size: usize) -> NonNull<u8> {
-    let non_zero_size = NonZeroUsize::new(size)
-        .unwrap_or_else(|| panic!("[Error] Allocation of zero bytes is not allowed."));
-    fluence::memory::alloc(non_zero_size)
-        .unwrap_or_else(|_| panic!("[Error] Allocation of {} bytes failed.", size))
-}
-
-/// Deallocates memory area for provided memory pointer and size.
-/// Used from the host environment for memory deallocation after reading results
-/// of function from Wasm memory.
-#[no_mangle]
-pub unsafe fn deallocate(ptr: NonNull<u8>, size: usize) {
-    let non_zero_size = NonZeroUsize::new(size)
-        .unwrap_or_else(|| panic!("[Error] Deallocation of zero bytes is not allowed."));
-    fluence::memory::dealloc(ptr, non_zero_size)
-        .unwrap_or_else(|_| panic!("[Error] Deallocate failed for ptr={:?} size={}.", ptr, size));
 }
 
 fn log_and_panic(msg: String) -> ! {
