+++ vm/examples/llamadb/Cargo.toml
@@ -1,10 +1,11 @@
 [package]
 
 name = "llama_db"
-version = "0.1.0"
+version = "0.1.1"
 authors = ["Fluence Labs"]
 publish = false
 description = "LlamaDb wrapper for running into Fluence WasmVm"
+edition = "2018"
 
 [profile.release]
 
@@ -23,9 +24,12 @@ crate-type = ["cdylib"]
 
 lazy_static = "1.1.0"
 log = "0.4"
-fluence_sdk = { version = "0.0.6", features = ["wasm_logger"] }
-llamadb = { git="https://github.com/fluencelabs/llamadb.git", branch = "master" }
+fluence = { version = "0.0.6", features = ["wasm_logger", "export_allocator"] }
+llamadb = { git = "https://github.com/fluencelabs/llamadb.git", branch = "master" }
 
 [dev-dependencies]
 
 simple_logger = "1.0.1"
+
+[replace]
+"fluence:0.0.6" = { path = "/Users/trofim/Desktop/work/fluence/fluence/sdk/rust" }
