+++ vm/examples/llamadb/src/lib.rs
@@ -27,20 +27,15 @@ mod tests;
 
 #[macro_use]
 extern crate lazy_static;
-#[macro_use]
-extern crate log;
-extern crate fluence_sdk as fluence;
-extern crate llamadb;
 
-use llamadb::tempdb::ExecuteStatementResponse;
-use llamadb::tempdb::TempDb;
+use log::{error, info, warn};
+use llamadb::tempdb::{ExecuteStatementResponse, TempDb};
 use std::error::Error;
-use std::num::NonZeroUsize;
 use std::ptr::NonNull;
 use std::sync::Mutex;
 
 /// Result for all possible Error types.
-type GenResult<T> = Result<T, Box<Error>>;
+type GenResult<T> = ::std::result::Result<T, Box<Error>>;
 
 //
 // FFI for interaction with Llamadb module.
@@ -84,39 +79,6 @@ pub unsafe fn invoke(ptr: *mut u8, len: usize) -> NonNull<u8> {
     // return pointer to result in memory
     fluence::memory::write_str_to_mem(&db_response)
         .unwrap_or_else(|_| log_and_panic("Putting result string to the memory was failed.".into()))
-}
-
-//
-// Public functions for memory management
-//
-
-/// Allocates memory area of specified size and returns its address.
-/// Used from the host environment for memory allocation while parameters passing.
-#[no_mangle]
-pub unsafe fn allocate(size: usize) -> NonNull<u8> {
-    info!("allocate starts with size={}", size);
-    let non_zero_size = NonZeroUsize::new(size).unwrap_or_else(|| {
-        log_and_panic("[Error] Allocation of zero bytes is not allowed.".into())
-    });
-    fluence::memory::alloc(non_zero_size)
-        .unwrap_or_else(|_| log_and_panic(format!("[Error] Allocation of {} bytes failed.", size)))
-}
-
-/// Deallocates memory area for provided memory pointer and size.
-/// Used from the host environment for memory deallocation after reading results
-/// of function from Wasm memory.
-#[no_mangle]
-pub unsafe fn deallocate(ptr: NonNull<u8>, size: usize) {
-    info!("deallocate starts with ptr={:?}, size={}", ptr, size);
-    let non_zero_size = NonZeroUsize::new(size).unwrap_or_else(|| {
-        log_and_panic("[Error] Deallocation of zero bytes is not allowed.".into())
-    });
-    fluence::memory::dealloc(ptr, non_zero_size).unwrap_or_else(|_| {
-        log_and_panic(format!(
-            "[Error] Deallocate failed for ptr={:?} size={}.",
-            ptr, size
-        ))
-    });
 }
 
 fn log_and_panic(msg: String) -> ! {
